+++ src/main/java/com/example/Dynamo_Backend/repository/CurrentStaffRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.Dynamo_Backend.entities.CurrentStaff;
 
-public interface CurrentStaffRepository extends JpaRepository<CurrentStaff, Integer> {
+public interface CurrentStaffRepository extends JpaRepository<CurrentStaff, Long> {
     CurrentStaff findByMachine_MachineId(Integer machineId);
 }
